// generated/src/app/api-models/contact-title-master.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ContactTitleMaster {
  contactTitle: string;
  description: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  titleType: string;
}